grayscaling/requirements.txt
boto3==1.17.109
botocore==1.20.109
cycler==0.10.0
decorator==4.4.2
imageio==2.9.0
jmespath==0.10.0
kiwisolver==1.3.1
matplotlib==3.4.2
networkx==2.5.1
numpy==1.21.0
Pillow==8.3.1
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
s3transfer==0.4.2
scipy==1.7.0
six==1.16.0
tifffile==2021.7.2
urllib3==1.26.6

debugpy>=1.0,<2